move test_getitem_duplicates_multiindex_empty_indexer to test_loc.py
