+++ pandas/tests/indexing/multiindex/test_getitem.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from pandas.compat import range, u, zip
+from pandas.compat import u, zip
 
 from pandas import DataFrame, Index, MultiIndex, Series
 import pandas.core.common as com
@@ -136,23 +136,6 @@ def test_series_getitem_corner_generator(
 # ----------------------------------------------------------------------------
 # test indexing of DataFrame with multi-level Index
 # ----------------------------------------------------------------------------
-
-@pytest.mark.parametrize('columns_indexer', [
-    ([], slice(None)),
-    (['foo'], [])
-])
-def test_getitem_duplicates_multiindex_empty_indexer(columns_indexer):
-    # GH 8737
-    # empty indexer
-    multi_index = MultiIndex.from_product((['foo', 'bar', 'baz'],
-                                           ['alpha', 'beta']))
-    df = DataFrame(np.random.randn(5, 6), index=range(5), columns=multi_index)
-    df = df.sort_index(level=0, axis=1)
-
-    expected = DataFrame(index=range(5), columns=multi_index.reindex([])[0])
-    result = df.loc[:, columns_indexer]
-    tm.assert_frame_equal(result, expected)
-
 
 def test_getitem_duplicates_multiindex_non_scalar_type_object():
     # regression from < 0.14.0
